requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
requests==2.32.3
fpdf==1.7.2
python-dotenv==1.0.1
